Update Readme for new version.
1.1.0 Release.

I decided to renumber the versions to give myself a bugfix version to
work on within each of the minor fixes. The new flow is this:

1.x: Introductory Rules Release
2.x: First of the Standard Rules Releases
3.x, etc: Subsequent Releases.

Things planned for subsequent releases include all the rules found in
the Alphastrike and Alphastrike Companion. I also want to include
Battleforce and Strategic Battleforce since those rules use the same
stat blocks, but that is some time away.