Separate out game logic from display logic
McCLIM is good, but I think I am going to need to switch to a more
portable GUI library. I'm considering GTK. In order to do this, I need
to cleanly separate the GUI code and the game logic.

This brings all the variables which were part of the application frame
out into two new classes: game and lobby. Additionally, it renames the
army class to force, in keeping with the game's naming convention.