GUI Cleanup
Trying to get the sizing to look better on the lobby screen.

Still not happy, I need a way to anchor the Not Ready button to the
bottom-right corner. I also have the problem that the map area is just
so small that there's no good place in the layout to put it.