Refactor the lobby window into a class
In the lobby, I need grids which are created by different functions to
be able to flag changes in other grids, created by other functions. This
was a particularly sticky wicket for adding units to the game. The
solution I had come up with worked so long as everything used the
string-list interface, but as soon as units were being drawn using the
list-box interface, it broke.

As an added benefit, I believe that this may also allow additional
refactoring to reduce the total count of lines in this file.